Omit <a> tags around content if URL is NULL.
It's common to have to write things like this, depending
on whether a button or link is enabled or not:

	{if $lang_switch_url}
		{a $lang_switch_url}
			<div class="language switch large button inner">
				<img src="images/double_arrows_around.jpg" />
				<span class="label">English / Local</span>
			</div>
		{/a}
	{else}
		<div class="language switch large button inner">
			<img src="images/double_arrows_around.jpg" />
			<span class="label">English / Local</span>
		</div>
	{/if}

To avoid the need to duplicate the stuff in the {a} block, we now omit
the <a> elements if the URL is NULL (and only then). But we still render
the contents of the block in both cases.